pyproject.toml
[project]
name = "flix-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-jazzmin>=3.0.1",
    "django>=5.1.3",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
]
